middleware module decomposed
- test script fix
- readme upd for nodejs
